Convert krunvm into a library
Convert krunvm into a library in order to enable its use as a crate in
other projects.
